finish implementation of EarthquakeAdapter

+++ 04 - Networking/ud843-QuakeReport-starting-point/app/src/main/java/com/example/android/quakereport/Earthquake.java
@@ -2,18 +2,18 @@ package com.example.android.quakereport;
 
 public class Earthquake
 {
-	private float mMagnitude;
+	private double mMagnitude;
 	private String mLocation;
 	private String mDate;
 
-	public Earthquake(float magnitude, String location, String date)
+	public Earthquake(double magnitude, String location, String date)
 	{
 		mMagnitude = magnitude;
 		mLocation = location;
 		mDate = date;
 	}
 
-	public float getMagnitude()
+	public double getMagnitude()
 	{
 		return mMagnitude;
 	}
